Changes to support files.
This is just a collection of meta-changes related to the rake tasks
for managing the Webby source code. Most notable is the removal of
svn tasks.

One minor change to the Webby webiste. Removed the "-y" flag from
the installation instructions. This flag is the default for the
latest version of RubyGems.